+++ src/Device/DeviceSelector/DeviceList/DeviceList.tsx
@@ -5,12 +5,17 @@
  */
 
 import React, { FC } from 'react';
-import { useSelector } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 import { bool, func } from 'prop-types';
 
 import { Device as DeviceProps } from '../../../state';
+import { Toggle } from '../../../Toggle/Toggle';
 import classNames from '../../../utils/classNames';
-import { sortedDevices } from '../../deviceSlice';
+import {
+    getGlobalAutoReconnect,
+    setGlobalAutoReconnect,
+    sortedDevices,
+} from '../../deviceSlice';
 import { AnimatedItem, AnimatedList } from './AnimatedList';
 import BrokenDevice from './BrokenDevice';
 import Device from './Device';
@@ -48,11 +53,23 @@ const DeviceList: FC<Props> = ({
     doSelectDevice,
     deviceFilter = showAllDevices,
 }) => {
+    const dispatch = useDispatch();
+    const autoReconnect = useSelector(getGlobalAutoReconnect);
     const devices = useSelector(sortedDevices);
     const filteredDevices = devices.filter(deviceFilter);
 
     return (
         <div className={classNames('device-list', isVisible || 'hidden')}>
+            <div className="global-auto-reconnect">
+                <Toggle
+                    id="toggle-global-auto-reconnect"
+                    label="Auto Reconnect"
+                    isToggled={autoReconnect}
+                    onToggle={value => {
+                        dispatch(setGlobalAutoReconnect(value));
+                    }}
+                />
+            </div>
             {devices.length === 0 && <NoDevicesConnected />}
             {devices.length > 0 && filteredDevices.length === 0 ? (
                 <NoSupportedDevicesConnected />