+++ src/state.ts
@@ -49,15 +49,21 @@ export interface Log {
     isLoggingVerbose: boolean;
 }
 
-export type Devices = { [key: string]: Device | undefined };
 export interface DeviceState {
-    devices: Devices;
+    devices: Map<string, Device>;
     deviceInfo: Device | null;
     isSetupDialogVisible: boolean;
     isSetupWaitingForUserInput: boolean | string;
     selectedSerialNumber: string | null;
     setupDialogChoices: readonly string[];
     setupDialogText?: string | null;
+    autoReconnectDevice?: AutoReconnectDevice | null;
+    autoReconnect: boolean;
+}
+
+export interface AutoReconnectDevice {
+    device: Device;
+    disconnectionTime?: number;
 }
 
 export interface DeviceInfo {
@@ -75,6 +81,7 @@ export interface Device extends NrfdlDevice {
     nickname?: string;
     serialport?: SerialPort;
     favorite?: boolean;
+    id: number;
 }
 
 export interface BrokenDeviceDialog {